Moved response decoders to a separate file
