+++ python_wrapper/susi_python/main.py
@@ -1,9 +1,8 @@
-import json
 import requests
 
-from susi_python.models import Answer, Datum, Metadata, Session, Identity, QueryResponse, LoginResponse, \
-    ForgotPasswordResponse, SignUpResponse, AnswerAction, TableAction, UnknownAction, Table, MapAction, AnchorAction, \
-    Map
+from susi_python.models import AnswerAction, TableAction, Table, MapAction, AnchorAction, Map
+from susi_python.response_decoders import MemoryResponseDecorder, ForgotPasswordDecoder, SignUpResponseDecoder, \
+    LoginResponseDecoder, SusiResponseDecoder
 
 api_endpoint = 'http://api.susi.ai'
 
@@ -75,81 +74,3 @@ def get_previous_responses():
     memory_url = api_endpoint + '/susi/memory.json'
     api_response = requests.get(memory_url)
     return api_response.json(cls=MemoryResponseDecorder)
-
-
-def get_action(jsn):
-    if jsn['type'] == 'answer':
-        return AnswerAction(jsn['expression'])
-    elif jsn['type'] == 'table':
-        return TableAction(jsn['columns'])
-    elif jsn['type'] == 'map':
-        return MapAction(jsn['latitude'], jsn['longitude'], jsn['zoom'])
-    elif jsn['type'] == 'anchor':
-        return AnchorAction(jsn['link'], jsn['text'])
-    else:
-        return UnknownAction()
-
-
-
-def generate_query_response(json_object):
-    ans = json_object['answers'][0]
-
-    data = [Datum(jsn)
-            for jsn in ans['data']]
-
-    metadata = Metadata(ans['metadata'])
-
-    actions = [get_action(jsn)
-               for jsn in ans['actions']]
-
-    answer = Answer(data, metadata, actions)
-
-    try:
-        identity_json = json_object['session']['identity']
-        identity = Identity(identity_json)
-        session = Session(identity)
-    except KeyError:
-        session = None
-
-    return QueryResponse(json_object, answer, session)
-
-
-class MemoryResponseDecorder(json.JSONDecoder):
-    def decode(self, raw_json):
-        json_object = super().decode(raw_json)
-        cognitions = json_object['cognitions']
-
-        susi_responses = []
-        for cognition in cognitions:
-            susi_responses.append(generate_query_response(cognition))
-
-        return susi_responses
-
-class SusiResponseDecoder(json.JSONDecoder):
-    def decode(self, raw_json):
-        json_object = super().decode(raw_json)
-        return generate_query_response(json_object)
-
-
-class LoginResponseDecoder(json.JSONDecoder):
-    def decode(self, raw_json):
-        json_object = super().decode(raw_json)
-        identity_json = json_object['session']['identity']
-        identity = Identity(identity_json)
-        session = Session(identity)
-        return LoginResponse(json_object, session)
-
-
-class ForgotPasswordDecoder(json.JSONDecoder):
-    def decode(self, raw_json):
-        json_object = super().decode(raw_json)
-        return ForgotPasswordResponse(json_object)
-
-
-class SignUpResponseDecoder(json.JSONDecoder):
-    def decode(self, raw_json):
-        json_object = super().decode(raw_json)
-        identity_json = json_object['session']['identity']
-        identity = Identity(identity_json)
-        session = Session(identity)
-        return SignUpResponse(json_object, session)
